chore: remove useClassNode customhook
